% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{main}
\alias{main}
\title{Convert OpenStreetMap OSM.PBF files into CSV files}
\usage{
main(filename)
}
\arguments{
\item{filename}{A character string; the path to the OSM.PBF file.}
}
\description{
Convert OpenStreetMap OSM.PBF files into CSV files
}
