% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosm4routing.R
\name{rosm4routing}
\alias{rosm4routing}
\title{Convert 'OpenStreetMap' PBF files to CSV files}
\usage{
rosm4routing(input, output = ".")
}
\arguments{
\item{input}{A character string; the input PBF file.}

\item{output}{A character string; the output directory.}
}
\description{
Convert 'OpenStreetMap' PBF files to CSV files
}
\examples{
\dontrun{
rosm4routing("Melbourne.osm.pbf")
}
}
